+++ src/Segues/ZoomOut.h
@@ -26,7 +26,7 @@ public:
     sf::Sprite left(*temp); 
     game::setOrigin(left, 0.5f, 0.5f);
     left.setPosition(windowSize.x/2.0f, windowSize.y/2.0f);
-    left.setScale(alpha, alpha);
+    left.setScale((float)alpha, (float)alpha);
 
     surface.clear();
